+++ tests/unit/base/test_api_config.py
@@ -31,6 +31,7 @@ from model_navigator.api.config import (
     _custom_configs,
     map_custom_configs,
 )
+from model_navigator.exceptions import ModelNavigatorConfigurationError
 
 
 def test_tensorflow_config_has_valid_name_and_format():
@@ -131,6 +132,23 @@ def test_map_custom_configs_return_dict_when_passed_list_of_configs():
     assert len(data) == 2
     assert data[tensorrt_config.name()] == tensorrt_config
     assert data[torch_tensorrt_config.name()] == torch_tensorrt_config
+
+
+def test_tensorrt_config_raise_error_when_invalid_optimization_level_provided():
+    with pytest.raises(
+        ModelNavigatorConfigurationError,
+        match="TensorRT `optimization_level` must be between 0 and 5. Provided value: 6.",
+    ):
+        TensorRTConfig(optimization_level=6)
+
+    with pytest.raises(
+        ModelNavigatorConfigurationError,
+        match="TensorRT `optimization_level` must be between 0 and 5. Provided value: -1.",
+    ):
+        TensorRTConfig(optimization_level=-1)
+
+    config = TensorRTConfig(optimization_level=2)
+    assert config.optimization_level == 2
 
 
 def test_custom_configs_raise_error_when_config_with_duplicated_name_defined():