+++ tests/unit/base/test_runtime_analyzer.py
@@ -34,6 +34,8 @@ tensorrt_config = TensorRTConfig(
     precision_mode=TensorRTPrecisionMode.HIERARCHY,
     max_workspace_size=DEFAULT_MAX_WORKSPACE_SIZE,
     trt_profile=None,
+    optimization_level=None,
+    compatibility_level=None,
 )
 
 model_statuses1 = {